Refactor Attribute Marshalers
Gets rid of the previous module-based system, in favor of marshaler
objects per attribute. When relevant, such as for Date/DateTime
marshalers, this also allows for attribute/marshaler level options such
as time zone offsets and formatting.
