--- lib/aws-record/record/attributes/string_marshaler.rb
@@ -1,60 +1,0 @@
-# Copyright 2015-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License"). You may not
-# use this file except in compliance with the License. A copy of the License is
-# located at
-#
-#     http://aws.amazon.com/apache2.0/
-#
-# or in the "license" file accompanying this file. This file is distributed on
-# an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-# or implied. See the License for the specific language governing permissions
-# and limitations under the License.
-
-module Aws
-  module Record
-    module Attributes
-      module StringMarshaler
-
-        class << self
-
-          def type_cast(raw_value, options = {})
-            case raw_value
-            when nil
-              if options[:nil_as_empty_string]
-                ''
-              else
-                nil
-              end
-            when String
-              if raw_value.empty? && !options[:nil_as_empty_string]
-                nil
-              else
-                raw_value
-              end
-            else
-              raw_value.to_s
-            end
-          end
-
-          def serialize(raw_value, options = {})
-             value = type_cast(raw_value)
-            if value.is_a?(String)
-              if value.empty?
-                nil
-              else
-                value
-              end
-            elsif value.nil?
-              nil
-            else
-              msg = "expected a String value or nil, got #{value.class}"
-              raise ArgumentError, msg
-            end
-          end
-
-        end
-      end
-    end
-  end
-end
